Made toString methods for meeting, Transactions. In transaction this is an abstract method that can be called on any type of transaction.
Made getTransactionItems as an abstract method that returns a list of items in the transaction (regardless of type of transaction.)

Made getTransactionMeetings as an abstract method that returns a list of meetings in the transaciton (regardless of type of transaction).

Deleted change meetingbystring methods.Added and checked over javadoc for transaction abstract and subclasses.