Made Transactions more extendable and re-implemented methods that are used around the program. Made extra Dcoumentation.
There was some sort of names dictionary for the items that got deleted and the item names in transactions aren't implemented because them being in the transaction doesn't make sense to me.

I know it's for the toString but do we need a toString in the transaction class? Why not make the string representation in a different class? (like usermenucontroller of have a facade design pattern).